requirements.txt
pandas
scikit-learn
joblib
azureml-core